+++ packages/clean_framework/example/lib/features/form/domain/form_entity.dart
@@ -7,7 +7,7 @@ enum FormTags {
   selectGender,
 }
 
-enum FormState {
+enum FormScreenState {
   initial,
   loading,
   success,
@@ -17,36 +17,37 @@ enum FormState {
 class FormEntity extends Entity {
   FormEntity({
     required this.formController,
-    this.state = FormState.initial,
-    this.userMeta = const UserMeta(),
+    this.screenState = FormScreenState.initial,
+    this.userMeta = const UserMetaEntity(),
     this.requireGender = true,
   });
 
   final FormController formController;
-  final FormState state;
-  final UserMeta userMeta;
+  final FormScreenState screenState;
+  final UserMetaEntity userMeta;
   final bool requireGender;
 
   @override
-  List<Object> get props => [formController, state, userMeta, requireGender];
+  List<Object> get props =>
+      [formController, screenState, userMeta, requireGender];
 
   @override
   FormEntity copyWith({
-    FormState? state,
-    UserMeta? userMeta,
+    FormScreenState? screenState,
+    UserMetaEntity? userMeta,
     bool? requireGender,
   }) {
     return FormEntity(
       formController: formController,
-      state: state ?? this.state,
+      screenState: screenState ?? this.screenState,
       userMeta: userMeta ?? this.userMeta,
       requireGender: requireGender ?? this.requireGender,
     );
   }
 }
 
-class UserMeta extends Entity {
-  const UserMeta({
+class UserMetaEntity extends Entity {
+  const UserMetaEntity({
     this.email = '',
     this.password = '',
     this.gender = '',