+++ packages/clean_framework/example/lib/features/form/domain/form_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:clean_framework/clean_framework.dart';
-import 'package:clean_framework_example/core/validators/validators.dart';
-import 'package:clean_framework_example/features/form/domain/form_entity.dart';
-import 'package:clean_framework_example/features/form/domain/form_ui_output.dart';
+import 'package:clean_framework_example_rest/core/validators/validators.dart';
+import 'package:clean_framework_example_rest/features/form/domain/form_entity.dart';
+import 'package:clean_framework_example_rest/features/form/domain/form_domain_models.dart';
 
 class FormUseCase extends UseCase<FormEntity> {
   FormUseCase()
@@ -11,7 +11,7 @@ class FormUseCase extends UseCase<FormEntity> {
               validators: {const InputFieldValidator.required()},
             ),
           ),
-          transformers: [FormUIOutputTransformer()],
+          transformers: [FormDomainToUIModelTransformer()],
         ) {
     _emailController = TextFieldController.create(
       entity.formController,
@@ -49,21 +49,22 @@ class FormUseCase extends UseCase<FormEntity> {
   Future<void> login() async {
     final formController = entity.formController;
     if (formController.validate()) {
-      entity = entity.copyWith(state: FormState.loading);
+      entity = entity.copyWith(screenState: FormScreenState.loading);
       formController.setSubmitted(true);
 
       // Simulates login
       await Future<void>.delayed(const Duration(seconds: 1));
 
-      final userMeta = UserMeta(
+      final userMeta = UserMetaEntity(
         email: _emailController.value ?? '',
         password: _passwordController.value ?? '',
         gender: _genderController.value?.name ?? '',
       );
-      entity = entity.copyWith(state: FormState.success, userMeta: userMeta);
+      entity = entity.copyWith(
+          screenState: FormScreenState.success, userMeta: userMeta);
       formController.setSubmitted(false);
 
-      entity = entity.copyWith(state: FormState.initial);
+      entity = entity.copyWith(screenState: FormScreenState.initial);
     }
   }
 
@@ -78,14 +79,14 @@ class FormUseCase extends UseCase<FormEntity> {
   }
 }
 
-class FormUIOutputTransformer
-    extends OutputTransformer<FormEntity, FormUIOutput> {
+class FormDomainToUIModelTransformer
+    extends DomainModelTransformer<FormEntity, FormDomainToUIModel> {
   @override
-  FormUIOutput transform(FormEntity entity) {
-    return FormUIOutput(
+  FormDomainToUIModel transform(FormEntity entity) {
+    return FormDomainToUIModel(
       formController: entity.formController,
-      isLoading: entity.state == FormState.loading,
-      isLoggedIn: entity.state == FormState.success,
+      isLoading: entity.screenState == FormScreenState.loading,
+      isLoggedIn: entity.screenState == FormScreenState.success,
       userMeta: entity.userMeta,
       requireGender: entity.requireGender,
     );