+++ packages/clean_framework/example/lib/features/home/presentation/home_presenter.dart
@@ -1,13 +1,13 @@
 import 'package:clean_framework/clean_framework.dart';
-import 'package:clean_framework_example/features/home/domain/home_entity.dart';
-import 'package:clean_framework_example/features/home/domain/home_ui_output.dart';
-import 'package:clean_framework_example/features/home/domain/home_use_case.dart';
-import 'package:clean_framework_example/features/home/presentation/home_view_model.dart';
-import 'package:clean_framework_example/providers.dart';
+import 'package:clean_framework_example_rest/features/home/domain/home_entity.dart';
+import 'package:clean_framework_example_rest/features/home/domain/home_domain_models.dart';
+import 'package:clean_framework_example_rest/features/home/domain/home_use_case.dart';
+import 'package:clean_framework_example_rest/features/home/presentation/home_view_model.dart';
+import 'package:clean_framework_example_rest/providers.dart';
 import 'package:flutter/material.dart';
 
 class HomePresenter
-    extends Presenter<HomeViewModel, HomeUIOutput, HomeUseCase> {
+    extends Presenter<HomeViewModel, HomeDomainToUIModel, HomeUseCase> {
   HomePresenter({
     required super.builder,
   }) : super(provider: homeUseCaseProvider);
@@ -18,26 +18,28 @@ class HomePresenter
   }
 
   @override
-  HomeViewModel createViewModel(HomeUseCase useCase, HomeUIOutput output) {
+  HomeViewModel createViewModel(
+      HomeUseCase useCase, HomeDomainToUIModel domainModel) {
     return HomeViewModel(
-      pokemons: output.pokemons,
-      onSearch: (query) => useCase.setInput(PokemonSearchInput(name: query)),
+      pokemons: domainModel.pokemons,
+      onSearch: (query) =>
+          useCase.setInput(PokemonSearchDomainInput(name: query)),
       onRefresh: () => useCase.fetchPokemons(isRefresh: true),
       onRetry: useCase.fetchPokemons,
-      isLoading: output.status == HomeStatus.loading,
-      hasFailedLoading: output.status == HomeStatus.failed,
-      loggedInEmail: output.loggedInEmail,
-      errorMessage: output.errorMessage,
+      isLoading: domainModel.status == HomeStatus.loading,
+      hasFailedLoading: domainModel.status == HomeStatus.failed,
+      loggedInEmail: domainModel.loggedInEmail,
+      errorMessage: domainModel.errorMessage,
     );
   }
 
   @override
-  void onOutputUpdate(BuildContext context, HomeUIOutput output) {
-    if (output.isRefresh) {
+  void onOutputUpdate(BuildContext context, HomeDomainToUIModel domainModel) {
+    if (domainModel.isRefresh) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
           content: Text(
-            output.status == HomeStatus.failed
+            domainModel.status == HomeStatus.failed
                 ? 'Sorry, failed refreshing pokemons!'
                 : 'Refreshed pokemons successfully!',
           ),