+++ packages/clean_framework/example/lib/features/profile/external_interface/pokemon_species_gateway.dart
@@ -1,41 +1,32 @@
 import 'package:clean_framework/clean_framework.dart';
-import 'package:clean_framework_example/core/pokemon/pokemon_request.dart';
-import 'package:clean_framework_example/core/pokemon/pokemon_success_response.dart';
-import 'package:clean_framework_example/features/profile/models/pokemon_species_model.dart';
+import 'package:clean_framework_example_rest/core/pokemon/pokemon_request.dart';
+import 'package:clean_framework_example_rest/core/pokemon/pokemon_success_response.dart';
+import 'package:clean_framework_example_rest/features/profile/domain/profile_domain_inputs.dart';
+import 'package:clean_framework_example_rest/features/profile/domain/profile_domain_models.dart';
+import 'package:clean_framework_example_rest/features/profile/models/pokemon_species_model.dart';
 
-class PokemonSpeciesGateway extends Gateway<PokemonSpeciesGatewayOutput,
-    PokemonSpeciesRequest, PokemonSuccessResponse, PokemonSpeciesSuccessInput> {
+class PokemonSpeciesGateway extends Gateway<
+    PokemonSpeciesDomainToGatewayModel,
+    PokemonSpeciesRequest,
+    PokemonSuccessResponse,
+    PokemonSpeciesSuccessDomainInput> {
   @override
-  PokemonSpeciesRequest buildRequest(PokemonSpeciesGatewayOutput output) {
-    return PokemonSpeciesRequest(name: output.name);
+  PokemonSpeciesRequest buildRequest(
+      PokemonSpeciesDomainToGatewayModel domainModel) {
+    return PokemonSpeciesRequest(name: domainModel.name);
   }
 
   @override
-  FailureInput onFailure(FailureResponse failureResponse) {
-    return FailureInput(message: failureResponse.message);
+  FailureDomainInput onFailure(FailureResponse failureResponse) {
+    return FailureDomainInput(message: failureResponse.message);
   }
 
   @override
-  PokemonSpeciesSuccessInput onSuccess(PokemonSuccessResponse response) {
-    return PokemonSpeciesSuccessInput(
+  PokemonSpeciesSuccessDomainInput onSuccess(PokemonSuccessResponse response) {
+    return PokemonSpeciesSuccessDomainInput(
       species: PokemonSpeciesModel.fromJson(response.data),
     );
   }
-}
-
-class PokemonSpeciesGatewayOutput extends Output {
-  PokemonSpeciesGatewayOutput({required this.name});
-
-  final String name;
-
-  @override
-  List<Object?> get props => [name];
-}
-
-class PokemonSpeciesSuccessInput extends SuccessInput {
-  PokemonSpeciesSuccessInput({required this.species});
-
-  final PokemonSpeciesModel species;
 }
 
 class PokemonSpeciesRequest extends GetPokemonRequest {