+++ packages/clean_framework/example/lib/routing.dart
@@ -1,2 +1,2 @@
-export 'package:clean_framework_example/routing/poke_router.dart';
-export 'package:clean_framework_example/routing/routes.dart';
+export 'package:clean_framework_example_rest/routing/poke_router.dart';
+export 'package:clean_framework_example_rest/routing/routes.dart';