+++ packages/clean_framework/example/test/widgets/svg_palette_card_test.dart
@@ -1,5 +1,5 @@
-import 'package:clean_framework_example/widgets/app_scope.dart';
-import 'package:clean_framework_example/widgets/svg_palette_card.dart';
+import 'package:clean_framework_example_rest/widgets/app_scope.dart';
+import 'package:clean_framework_example_rest/widgets/svg_palette_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:palette_generator/palette_generator.dart';