+++ packages/clean_framework/theme_example/lib/features/home/presentation/theme/home_theme_view_model.dart
@@ -1,0 +1,15 @@
+import 'package:clean_framework/clean_framework.dart';
+import 'package:theme_example/features/home/domain/home_entity.dart';
+
+class HomeThemeViewModel extends ViewModel {
+  const HomeThemeViewModel({
+    required this.appTheme,
+  });
+
+  final AppTheme appTheme;
+
+  @override
+  List<Object> get props => [
+        appTheme,
+      ];
+}