+++ packages/clean_framework_firestore/CHANGELOG.md
@@ -1,4 +1,16 @@
 # Changelog
+## 0.5.3
+**Feb 10, 2024**
+- Fix bug causing `FirebaseDeleteRequest`s to fail.
+
+## 0.5.2
+**Feb 7, 2024**
+- Upgrade package to use newest clean_framework classes.
+
+## 0.5.2-dev.0
+**Jan 29, 2024**
+- Bump dependencies to latest versions.
+
 ## 0.5.1
 **Aug 7, 2023**
 - Bumps dependencies to their latest version.